saju-project/requirements.txt
langchain
openai
langchain-openai
bazica
python-dotenv
